vfs_ceph_new: fix typecast error in vfs_ceph_flistxattr

Autobuild-User(master): Anoop C S <anoopcs@samba.org>
Autobuild-Date(master): Fri Mar 20 06:35:12 UTC 2026 on atb-devel-224

(cherry picked from commit cecbdb73fbf103d68d0275adaf8083a730171fce)

+++ source3/modules/vfs_ceph_new.c
@@ -4355,7 +4355,7 @@ static ssize_t vfs_ceph_flistxattr(struct vfs_handle_struct *handle,
 			goto out;
 		}
 	}
-	ret = (int)list_size;
+	ret = (ssize_t)list_size;
 out:
 	DBG_DEBUG("[CEPH] flistxattr done: ret=%zd\n", ret);
 	END_PROFILE_X(syscall_flistxattr);
